Module tnb_rs::nodes

Different nodes on the network

Structs

ConfirmationValidator

Confirmation Validator

PrimaryValidator

Primary Validator

RegularNode

A regular node that retrieves chain data and forwards blocks to the primary validator

Traits

ServerNodeTrait

A trait for implementing nodes using the server node as a base layer.

ValidatorTrait

A trait for implementing validator nodes using the server node as a base layer.