Struct tnb_rs::Account

pub struct Account { /* fields omitted */ }

An Account consists of an account number and a signing key.

The account number is used to identify an account on thenewboston network

The signing key is used to prove that you are the owner of the account by creating signatures and validating them with your account number. For this reason, your signing key should not be shared with anyone.

Implementations

impl Account

pub fn new() -> Self

Creates an random Account

    use tnb_rs::Account;

    let acc = Account::new();

    // store you signing key safely before you proceed
    println!("siging key: {}", acc.signing_key());

pub fn from_signing_key(signing_key_hex: &str) -> Result<Self>

Creates an Account from a signing key

Example

    use tnb_rs::Account;

    let signing_key = "8cf08eb96b00b5a4df86a750bb7ae595a9dbbe91fc091463bfb3d950d5dac467";
    let acc = Account::from_signing_key(signing_key).unwrap();

    assert_eq!(acc.signing_key(), signing_key);

pub fn account_number(&self) -> &str

Returns the account number as a hex string

pub fn signing_key(&self) -> &str

Returns the signing key as a hex string

pub fn keypair_as_hex(&self) -> (&str, &str)

Returns a tuple of the account number and signing key

pub fn is_valid_keypair(
_signing_key_hex: &str,
_account_number_hex: &str
) -> bool

Checks if a signing key and account number are keypairs

Example

     
    use tnb_rs::Account;

    let signing_key = "8cf08eb96b00b5a4df86a750bb7ae595a9dbbe91fc091463bfb3d950d5dac467";
    let acc = Account::from_signing_key(signing_key).unwrap();
    let (acc_num, sk ) = acc.keypair_as_hex();
     
    assert_eq!(signing_key, sk);

    let is_keypair = Account::is_valid_keypair(&sk, &acc_num);
    assert_eq!(is_keypair, true);
     

pub fn create_signature(&self, message: &str) -> String

Signs the given message with the Account’s signing key

Returns

A signature of the signed message in form of a hex string.

pub fn verify_signature(
signature_as_hex: &str,
message: &str,
account_number_hex: &str
) -> bool

Static method that the verifies that a message was signed by a specific account number.

Inputs

signature being verified

the original message

the signer’s account number

Returns

A signature in form of a hex string

pub fn create_block_message<'a>(
&'a self,
data: &'a BlockData<'_>
) -> BlockMessage<'_>

Creates a block message that can be broadcasted to make changes to an account on the network

pub fn create_signed_message<'a>(
&'a self,
data: &'a ChainData
) -> SignedMessage<'_>

Creates a message that nodes can broadcast to make changes on the network