Struct tnb_rs::HDWallet

pub struct HDWallet {
    pub mnemonic: String,
    // some fields omitted
}

A Hierarchical Deterministic Wallet that can create an infinite number of Accounts from a single mnemonic phrase.

To specify the location of accounts in the tree we use something called path a path

This wallet uses the bip44 standard for compatibility with other existing wallets Therefor all paths are formatted like this path format:

m / 44' / coin_type' / account_index' / 0' / address_index'

The coin_type is specific to the cryptocurrency network and thenewboston’s is 2002

so the paths are

    m / 44' / 2002' / account_index' / 0' / address_index'

The account_index and address_index can be changed by the user as they please to get their desired account

With this format a user can still get up to 4,611,686,014,132,420,600 accounts from a single mnemonic phrase

Fields
mnemonic: String

mnemonic phrase

Implementations

impl HDWallet

pub fn new() -> Self

Creates a new HD Wallet

    use tnb_rs::HDWallet;
    let hd = HDWallet::new();
    println!("mnemonic: {}", hd.mnemonic);

Ensure that you save your mnemonic phrase somewhere secure so you can use it again later

pub fn from_mnemonic(mnemonic: &str, password: Option<&str>) -> Self

Creates a HD Wallet from a specified mnemonic phrase.

Inputs

mnemonic phrase - a 12, 18, or 24 random words generated from a wordlist

password - an optional string to further secure your wallet in case your mnemonic phrase is ever exposed

Example

    use tnb_rs::HDWallet;

    let mnemonic = "visa nephew like this amazing soldier negative front elevator warfare teach good";
    let hd1 = HDWallet::from_mnemonic(mnemonic, None);

    let hd2 = HDWallet::from_mnemonic(mnemonic, Some("password"));

    // same mnemonic phrase
    assert_eq!(hd1.mnemonic, hd2.mnemonic);

    // but different seeds meaning all the accounts generated will be different too.   
    assert_ne!(hd1.seed_hex(), hd2.seed_hex());

pub fn new_with_password(password: Option<&str>) -> Self

Creates a new HD Wallet with an optional field for a password

pub fn from_seed(seed: String) -> Self

Creates a HD Wallet from a seed in hex string format

pub fn get_account(&self, account_index: u32, address_index: u32) -> Account

Retrieves the account specified by the account_index and address_index from the HD Wallet

Example

    use tnb_rs::{HDWallet, MAX_CHILD_INDEX};
     
    let mnemonic = "visa nephew like this amazing soldier negative front elevator warfare teach good";
    let hd = HDWallet::from_mnemonic(mnemonic, None);
     
    let acc1 = hd.get_account(0, 0);
    let acc2 = hd.get_account(23, 1_000);

    let last_account = hd.get_account(MAX_CHILD_INDEX, MAX_CHILD_INDEX);

pub fn get_account_0(&self, address_index: u32) -> Account

Retrieves an account specified by the address_index from the column where (account_index == 0)

pub fn get_first_account(&self) -> Account

Retrieves the first account from the hd wallet

Example

    use tnb_rs::HDWallet;
     
    let mnemonic = "visa nephew like this amazing soldier negative front elevator warfare teach good";
    let hd = HDWallet::from_mnemonic(mnemonic, None);

    let get_first_account = hd.get_first_account();

    assert_eq!(get_first_account.account_number_hex(),
                hd.get_account(0, 0).account_number_hex());
    assert_eq!(get_first_account.signing_key_hex(),
                hd.get_account(0, 0).signing_key_hex());

pub fn seed_hex(&self) -> String

Returns the seed in hex format